CLI: include META-INF/services/ from kotlin-reflect with -include-runtime
This is an addition to bd205317aae15816783df332eae851483167f3cd where we
started to add the contents of kotlin-reflect into the resulting jar if
-include-runtime is specified. Apparently kotlin-reflect doesn't work
without some services in META-INF/services/, which didn't satisfy the
condition for inclusion. The existing test didn't catch that because
loading class annotations does not always lead to those services being
loaded.